% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daysAboveThreshold.R
\name{daysAboveThreshold}
\alias{daysAboveThreshold}
\title{Detect days above a threshold}
\usage{
daysAboveThreshold(data, threshold, type = NULL)
}
\arguments{
\item{data}{A data frame containing time series data.}

\item{threshold}{A numeric to calculate how long is exceeded}

\item{type}{A string. "rising" or "falling" (or NULL) i.e. "rising" will only accumulate times were level is higher than last level}
}
\value{
A vector (in days)
}
\description{
This function calculates the cumulative number of days that a sensor value exceeds a specified threshold.
It can also be configured to only accumulate time when the sensor value is rising or falling, based on the `type` parameter.
}
\note{
The input data frame (`data`) must include:

| Column Name | Type      | Description                                       |
|-------------|-----------|---------------------------------------------------|
| `<time_col>` | POSIXct   | Timestamp of each observation                     |
| `<value_col>`| numeric   | Measured sensor value                             |

All other columns in the input data frame will be preserved in the output.
}
\examples{
library(dplyr)
# Simulated data
set.seed(42)
ts <- seq.POSIXt(from = as.POSIXct("2024-01-01"), by = "hour", length.out = 200)
val <- c(rnorm(150, mean = 2), rep(6, 50))  # final 50 points simulate drift
df <- data.frame(ts = ts, Value = val, Quality = NA_character_)
daysAbove <- daysAboveThreshold(df, 2)

plot(df$ts, daysAbove)

}
