% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_detect_rt.R
\name{drift_detect_rt}
\alias{drift_detect_rt}
\title{Detect gradual sensor drift in time series water quality data}
\usage{
drift_detect_rt(
  ts,
  value,
  last_t,
  last_value,
  threshold_multiplier = 1.1,
  time_threshold_days = 2,
  halflife = 120,
  type = NULL,
  last_mean = NULL,
  cumulative_time = NULL
)
}
\arguments{
\item{ts}{POSIXCt This is the timestamp of the current observation.}

\item{value}{Numeric This is the sensor value of the current observation.}

\item{last_t}{POSIXCt This is the timestamp of the last observation.}

\item{last_value}{Numeric This is the sensor value of the last observation.}

\item{threshold_multiplier}{A numeric multiplier applied to the moving average value to define the drift threshold. Default is 1.1.}

\item{time_threshold_days}{Minimum number of **continuous** days above the threshold required to flag sensor drift. Default is 2 days.}

\item{halflife}{Numeric The halflife (in days) for the exponential decay used in calculating the dynamic mean. Default is 120 days.}

\item{type}{A string indicating the type of drift to detect: "rising", "falling", or NULL for both. Default is NULL.}

\item{last_mean}{Numeric The mean value from the previous observation, used for dynamic threshold}

\item{cumulative_time}{Numeric The cumulative time (in days) above the threshold from the previous observation.}
}
\value{
A data frame with the input time and value, plus recursive parameters to feed back into the function
\describe{
\item{ts}{Timestamp of the observation}
\item{value}{Observed value}
\item{cumulative_time}{Cumulative time above threshold}
\item{mean}{Updated running exponential time weighted mean}
\item{drift}{Logical stating whether or not the point is drift}
\item{z_score}{Z-score of the current observation based on the running mean
and variance}
\item{spike}{Logical indicator of whether the current observation is classified as a spike (TRUE) or not (FALSE) based on the z-score thresholds}
}
}
\description{
This function detects sustained periods where sensor readings remain consistently
above a calculated threshold, which may indicate sensor drift. The threshold is
defined as a multiple of the running mean, and drift is flagged if readings exceed
this threshold continuously for longer than a specified number of days.
}
\details{
Depending on the value of "type", the function will accumulate either rising or falling points, or both.
}
\examples{
library(dplyr)
library(data.table)
library(ggplot2)
# Simulated data
set.seed(42)
ts <- seq.POSIXt(from = as.POSIXct("2024-01-01"), by = "hour", length.out = 200)
val <- c(rnorm(150, mean = 2), rep(6, 50))  # final 50 points simulate drift
df <- data.frame(ts = ts, Value = val, Quality = NA_character_)

drift_results <- list()
for(i in 1:nrow(df)){
  if (i == 1) {
    drift_df <- drift_detect_rt(ts = df[i,]$ts, value = df[i,]$Value,
      last_t = 0, last_value = 0)
  }else{
    drift_df <- drift_detect_rt(ts= df[i,]$ts, value = df[i,]$Value,
      last_t = drift_df$ts, last_value = drift_df$value,
      last_mean = drift_df$mean, cumulative_time = drift_df$cumulative_time
      )
  }
  drift_results[[length(drift_results) + 1]] <- drift_df
}
drift_results <- rbindlist(drift_results)

drift_results \%>\% ggplot(aes(x = ts, y = value, colour = drift)) +
geom_point()

\dontrun{
# Visualize
if (interactive() && requireNamespace("plotly", quietly = TRUE)) {
  plotly::plot_ly(drift_results) |>
    plotly::add_markers(
      x = ~ts,
      y = ~value,
      type = "scatter",
      color = ~drift
    )
}
}

}
