% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagleIO.R
\name{eio_hist}
\alias{eio_hist}
\title{Extract historic data from WQI's eagle.IO instance}
\source{
\url{https://wqi.eagle.io/}
}
\usage{
eio_hist(key, node_id, start_time = NULL, end_time = Sys.time())
}
\arguments{
\item{key}{This is required for all functions and can be generated from the account settings in WQI's Eagle.IO instance}

\item{node_id}{A node ID of a given parameter in eagle.IO, noting historic will only work with nodes that contain historic data (ie level/N-NO3/Turbidity).}

\item{start_time}{DateTime to start the historic period of capture. Format needs to POSIXCt or in ISO8601. If it is POSIXCt it will be converted to correctly to UTC. If it is ISO8601 it will be used as is and will be treated as already in UTC.}

\item{end_time}{DateTime to end the historic period of capture. Defaults to the current system time + one day. Format needs to be in ISO8601, see: 2014-10-09T22:38:10Z || 2014-10-09T22:38:10.000Z || 2014-10-09T20:38:10+0200 || 2014-10-09T20:38:10+02:00}
}
\value{
tibble containing returned historic data, value and quality. Time as "Australia/Brisbane"
}
\description{
Function for extracting historic data from WQI's eagle.IO instance. Requires a node ID that has connection status information to return (ie. Campbell pakbus logger)
}
\examples{
#td <- 86400
#START <- format(Sys.time() -30*td, "\%Y-\%m-\%dT\%H:\%M:\%SZ") # note this will be in UTC
#content <- eio_hist(key = "XYZ", param = "5903e538bd10c2fa0ce50648", START = 1)
#library(tidyverse)
#reportableParamRef <- WQI::reportableParamRef
#param <- reportableParamRef |>
#  filter(GSnum == '1160122')
#param <- param$`N-NO3`

}
