% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{LongPaddock_grid}
\alias{LongPaddock_grid}
\title{Retrieve gridded data from LongPaddock}
\usage{
LongPaddock_grid(
  lat,
  long,
  start,
  finish = format(Sys.time() + 86400, "\%Y\%m\%d"),
  format = "standard",
  username
)
}
\arguments{
\item{lat}{The latitude.}

\item{long}{The longitude.}

\item{start}{The start date in the format "YYYYMMDD".}

\item{finish}{The finish date in the format "YYYYMMDD".}

\item{format}{The format of the data. The default is "standard".}

\item{username}{The username for LongPaddock.}
}
\value{
A tibble with the gridded data.
}
\description{
This function retrieves gridded data from LongPaddock. The function takes
the latitude, longitude, start date, finish date, format, and username as input. The
function returns a tibble with the gridded data.
}
\examples{
df <- LongPaddock_grid(lat = "-18.50",
                      long = "145.98",
                     start = format(Sys.time() - (700*86400), "\%Y\%m\%d"),
                     format = "alldata",
                     username = "example@email.com.au")

}
\seealso{
https://www.longpaddock.qld.gov.au/cgi-bin/silo/DataDrillDataset.php
}
