% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{plt_rain}
\alias{plt_rain}
\title{Plot rainfall data}
\usage{
plt_rain(df)
}
\arguments{
\item{df}{A dataframe with rainfall data.}
}
\value{
A Plotly figure.
}
\description{
This function plots rainfall data using Plotly. The function takes
a dataframe with rainfall data as input. The
function returns a Plotly figure.
}
\examples{
df <- LP_nearSite(station = "32002")

fig <- plt_rain(df)

}
\seealso{
https://plotly.com/r/
}
