% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{silo_nearby}
\alias{silo_nearby}
\title{Retrieve nearby-site information from LongPaddock}
\usage{
silo_nearby(station, radius = "50")
}
\arguments{
\item{station}{The station ID.}

\item{radius}{The radius in kilometers.}
}
\value{
A tibble with the near-site data.
}
\description{
This function retrieves nearby-site information from LongPaddock. The function takes
the station ID and radius as input. The
function returns a tibble with the near-site data.
}
\examples{
df <- silo_nearby(station = "32002")

}
\seealso{
https://www.longpaddock.qld.gov.au/cgi-bin/silo/PatchedPointDataset.php?format=near
}
