% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_detect_rt.R
\name{spike_detect_rt}
\alias{spike_detect_rt}
\title{Detect anomalies in time series water quality or quantity data}
\usage{
spike_detect_rt(
  ts,
  value,
  last_t = 0,
  last_running_mean = NULL,
  last_running_var = NULL,
  last_dev = NULL,
  halflife_mins = 120,
  z_threshold = c(0.001, 3),
  sd_floor = 0.01
)
}
\arguments{
\item{ts}{Timestamp of class `POSIXct`}

\item{value}{Numeric value}

\item{last_t}{Last timestamp processed (for real-time processing)}

\item{last_running_mean}{Last running mean value (for real-time processing)}

\item{last_running_var}{Last running variance value (for real-time processing)}

\item{last_dev}{Last deviation value (for real-time processing)}

\item{halflife_mins}{Half-life in minutes for exponential decay in running mean. The lower the number, the more responsive to real world changes. However shortens "memory".}

\item{z_threshold}{Numeric vector of length 2; lower and upper thresholds for z-score to classify spikes. Values outside this range are considered spikes. A lower upper z_score will narrow the bands around the data, detecting more spikes - but may incorrectly flag real-world changes.}

\item{sd_floor}{Numeric a higher number reduces the number "noise" spikes.

# make the return describe each part and what it means,}
}
\value{
A `data.frame` with output fields serving as the updated state of the running statistics after processing the current observation:
\describe{
\item{ts}{Timestamp of the observation}
\item{value}{Observed value}
\item{mean}{Updated running mean after processing the current observation}
\item{var}{Updated running variance after processing the current observation}
\item{dev}{Deviation of the current observation from the running mean}
\item{z_score}{Z-score of the current observation based on the running mean
and variance}
\item{spike}{Logical indicator of whether the current observation is classified as a spike (TRUE) or not (FALSE) based on the z-score thresholds}
}
}
\description{
This function identifies anomalies in time series data based on:
- Values outside expected physical sensor bounds (`sensorMin`/`sensorMax`)
- Impossible values (e.g., negative values for non-negative sensors)
- Repeated or flatlined values using a rolling standard deviation
- Spikes using deviation from rolling median
}
\details{
Anomalies are flagged in a "Quality" column (or similarly named column based on what's present in the data).
}
\examples{
library(dplyr)
library(zoo)
library(plotly)
library(data.table)
library(ggplot2)

#example Total Suspended Solids dataframe
df <- waterQUAC::TSS_data

# Apply anomaly detection
flagged <- spike_detect_rt(ts = df$ts[1],
                         value = df$Value[1]
                         )
print(flagged)
# use last values as a recursive input to the functino
flagged <- spike_detect_rt(ts = df$ts[2],
                        value = df$Value[2],
                        last_t = flagged$ts,
                        last_running_mean = flagged$mean,
                        last_running_var = flagged$var,
                        last_dev = flagged$dev
                        )
print(flagged)

# loop through entire example df using a for loop
spike_results <- list()
for(i in 1:nrow(df)) {
  if(i == 1) {
    spike_results[[i]] <- spike_detect_rt(ts = df$ts[i],
                                     value = df$Value[i])
   }else{
      spike_results[[i]] <- spike_detect_rt(ts = df$ts[i],
                                            value = df$Value[i],
                                            last_t = spike_results[[i-1]]$ts,
                                            last_running_mean = spike_results[[i-1]]$mean,
                                            last_running_var = spike_results[[i-1]]$var,
                                           last_dev = spike_results[[i-1]]$dev
                                            )
   }
}
spike_results <- data.table::rbindlist(spike_results)

# and then add the upper and lower to the plotly plot:
spike_results <- spike_results \%>\%
mutate(upper = mean + 4* pmax(sqrt(var),0.01)) \%>\%
mutate(lower = mean - 4* pmax(sqrt(var),0.01))

head(spike_results)
spike_results \%>\% dplyr::select(-c(var, dev, z_score, mean)) \%>\%
  ggplot(.) +
  geom_point(aes(x=ts, y=value, colour = spike, shape = spike)) +
  geom_line(aes(x=ts, y=upper), colour = "grey", linetype = "dashed") +
  geom_line(aes(x=ts, y=lower), colour = "grey", linetype = "dashed")




\dontrun{
# Plot flagged results
  plotly::plot_ly(data = spike_results) |>
    plotly::add_markers(
      x = ~ts,
      y = ~value,
      color = ~spike,
      marker = list(size = 8)
    ) \%>\%
    plotly::add_lines(x = ~ts, y = ~upper, name = "Upper Threshold",
      line = list(color = 'grey', dash = 'dot')) \%>\%
    plotly::add_lines(x = ~ts, y = ~lower, name = "Lower Threshold",
      line = list(color = 'grey', dash = 'dot'))
 }
}
