% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathFinder.R
\name{trios_path}
\alias{trios_path}
\title{Decision tree for Trios OPUS and NICO Path-length limits for N-NO3 and TSSeq}
\source{
\url{https://www.trios.de/en/}
}
\usage{
trios_path(probe, path)
}
\arguments{
\item{probe}{The type of probe (OPUS/NICO)}

\item{path}{the path-length of interest for the relevant probe (2/5/10/20mm)}
}
\value{
List of the relevant limit values for given probe and pathlength
}
\description{
This function, given a valid path-length and probe type will return a list of the relevant limit values for available parameters
}
\examples{
#Limits <- PathFinder(ProbeType = "OPUS", PathLength = 5)
#Limits$NO3LowerLim

}
