% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnom.R
\name{ts.anom}
\alias{ts.anom}
\title{Detect anomalies in time series data}
\usage{
ts.anom(df, overwrite, sensorMin, sensorMax, window = 3, prec = 0.001)
}
\arguments{
\item{df}{A data frame containing the time series water quality or quantity data.}

\item{overwrite}{A vector of quality codes that CAN be overwritten by the function.}

\item{sensorMin}{The minimum value as reportable from the sensor for the given data.}

\item{sensorMax}{The maximum value as reportable from the sensor for the given data.}

\item{window}{The number of hours the average data logging interval is used to calculate the rolling window for flatline detection. For example if the average read time is hourly, flatln = 3 means consecutive values across a 3 hour interval will flag.}

\item{prec}{The precision of the sensor OR how much variation is acceptable to determine a repeating value}
}
\value{
A data frame with the original data and a new column called Quality that indicates the type of anomaly, if any.
}
\description{
This function detects anomalies in time series data by looking for values that are
outside of the expected range, or that are repeated values. The function can also
detect spikes in the data.
}
\examples{

df <- waterQUAC::TSS_data

manual_codes = c(1:4000)
sensorMin = 0
sensorMax = 650

tst <- ts.anom(df = df,
              overwrite = manual_codes,
              sensorMin = 0,
              sensorMax = 650)
tst |>
 plotly::plot_ly() |>
 plotly::add_markers(
   x =  ~ ts,
   y =  ~ Value,
   type = "scatter",
   color = ~ Quality
 )

}
