% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSS_data}
\alias{TSS_data}
\title{An example time series dataset of TSS probe data.}
\format{
a data frame containing 66851 obs. of  3 variables
\describe{
  \item{ts}{ts}
  \item{Value}{Value}
  \item{Quality}{Quality}
  }
}
\usage{
TSS_data
}
\description{
An example time series dataset of TSS probe data.
}
\keyword{datasets}
