% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTSDuplicates.R
\name{removeTSDuplicates}
\alias{removeTSDuplicates}
\title{Detect duplicated values in a timeseries}
\usage{
removeTSDuplicates(
  originalTS,
  diffColumn,
  maxdupes = 3,
  prec = 1e-04,
  output = 0
)
}
\arguments{
\item{originalTS}{A data frame containing time series data. It must have a timestamp column of class `POSIXct` and a numeric value column (assumed to be the second column).}

\item{diffColumn}{A vector of quality codes that CAN be overwritten by this function (e.g., manually applied flags that can be updated).}

\item{maxdupes}{Numeric, Number of duplicates to flag as repeating value}

\item{prec}{Numeric, similarity of a number to be a duplicate}

\item{output}{Integer, 0 (output the clean data) or 1 (output the duplicates)}
}
\value{
A dataframe with duplicates values removed
}
\description{
Detect duplicated values in a timeseries
}
\note{
The input data frame (`df`) must include the following columns:

| Column Name | Type      | Description                                               |
|-------------|-----------|-----------------------------------------------------------|
| `ts`        | POSIXct   | Timestamp of each observation                             |
| `Value`     | numeric   | Measured value from the sensor                            |

All other columns in the input data frame will be preserved in the output.
If `diag = TRUE`, additional diagnostic columns (e.g., rolling SD, median) will be appended.
}
