% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsAnom.R
\name{ts.anom}
\alias{ts.anom}
\title{Detect anomalies in time series data}
\usage{
ts.anom(
  df,
  OVERWRITE,
  sensorMin,
  sensorMax,
  flatln = 3,
  med_width = 36,
  sd_width = 100
)
}
\arguments{
\item{df}{A data frame containing the time series water quality or quantity data.}

\item{OVERWRITE}{A vector of quality codes that can be overwritten by the function.}

\item{sensorMin}{The minimum value as reportable from the sensor for the given data.}

\item{sensorMax}{The maximum value as reportable from the sensor for the given data.}

\item{flatln}{The number of times the average data logging interval is used to calculate the rolling window for flatline detection. For example if the average read time is hourly, flatln = 3 means consecutive values across a 3 hour interval will flag.}

\item{med_width}{The width of the rolling window used to calculate the median of the log(value).}

\item{sd_width}{The width of the rolling window used to calculate the standard deviation of the median log(value).}
}
\value{
A data frame with the original data and a new column called Quality that indicates the type of anomaly, if any.
}
\description{
This function detects anomalies in time series data by looking for values that are
outside of the expected range, or that are repeated values. The function can also
detect spikes in the data.
}
\examples{

#OVERWRITE <-
#  c(155,
#    101,
#    105,
#    2200,
#    1020,
#    2000,
#    2010:2015,
#    2210:2215,
#    3061:3068,
#    3211:3214) #overwritable QC codes, all else are retained
#
#
#
#tst <- ts.anom(df, OVERWRITE, 0, 650)
#
#tst \%>\%
#  plotly::plot_ly() \%>\%
#  plotly::add_markers(x=~ts, y=~Value, type = "scatter", color = ~Quality)

}
